Revert "Extract class CacheController"
This reverts commit ec8f88d292a58c5195b7b3ec9d748eb564f72886.

That commit was a really bad idea.  Most notably, it broke the basic
functionality of the plugin, because the model was no longer properly
initialized in LanguageMenuController and MainAdminController.  In
addition, the CacheController should not be a controller at all, and
of course, injecting the model as dependency to the controllers is
actually a Good Thing, anyway.  Not sure, was I was thinking when I
made that commit; I might have dreamed of an automated test suite.